requirements.txt
fastapi
uvicorn
python-multipart
jinja2
requests
yt-dlp
google-generativeai
python-dotenv
pydub
openai-whisper
torch
deep-translator
